refactor: join user controllers 🔥
